FS-5062 - Transferring tests of helpers module from general test_config_export into separate test module
