+++ tests/unit/app/export_config/test_config_export.py
@@ -11,7 +11,6 @@ from app.export_config.generate_fund_round_html import (
     frontend_html_suffix,
     generate_all_round_html,
 )
-from app.export_config.helpers import validate_json
 
 
 def test_generate_fund_round_html(seed_dynamic_data, temp_output_dir):
@@ -48,36 +47,6 @@ def test_generate_fund_round_html_invalid_input(seed_dynamic_data):
     # Execute and Assert: Ensure the function raises an exception for invalid inputs
     with pytest.raises(ValueError):
         generate_all_round_html(round_id)
-
-
-test_json_schema = {
-    "type": "object",
-    "properties": {"name": {"type": "string"}, "age": {"type": "number"}},
-    "required": ["name", "age"],
-}
-
-
-def test_valid_data_validate_json():
-    # Data that matches the schema
-    data = {"name": "John Doe", "age": 30}
-    result = validate_json(data, test_json_schema)
-    assert result, "The data should be valid according to the schema"
-
-
-@pytest.mark.parametrize(
-    "data",
-    [
-        ({"age": 30}),  # Missing 'name'
-        ({"name": 123}),  # 'name' should be a string
-        ({"name": ""}),  # 'name' is empty
-        ({}),  # Empty object
-        ({"name": "John Doe", "extra_field": "not allowed"}),  # Extra field not defined in schema
-        # Add more invalid cases as needed
-    ],
-)
-def test_invalid_data_validate_json(data):
-    result = validate_json(data, test_json_schema)
-    assert not result, "The data should be invalid according to the schema"
 
 
 def test_create_export_zip(temp_output_dir):
