+++ tests/unit/app/export_config/test_helpers.py
@@ -1,0 +1,32 @@
+import pytest
+
+from app.export_config.helpers import validate_json
+
+test_json_schema = {
+    "type": "object",
+    "properties": {"name": {"type": "string"}, "age": {"type": "number"}},
+    "required": ["name", "age"],
+}
+
+
+def test_valid_data_validate_json():
+    # Data that matches the schema
+    data = {"name": "John Doe", "age": 30}
+    result = validate_json(data, test_json_schema)
+    assert result, "The data should be valid according to the schema"
+
+
+@pytest.mark.parametrize(
+    "data",
+    [
+        ({"age": 30}),  # Missing 'name'
+        ({"name": 123}),  # 'name' should be a string
+        ({"name": ""}),  # 'name' is empty
+        ({}),  # Empty object
+        ({"name": "John Doe", "extra_field": "not allowed"}),  # Extra field not defined in schema
+        # Add more invalid cases as needed
+    ],
+)
+def test_invalid_data_validate_json(data):
+    result = validate_json(data, test_json_schema)
+    assert not result, "The data should be invalid according to the schema"
